// Backend/src/config.d.ts
import "dotenv/config";
export declare const config: {
    DATABASE_URL: string;
    JWT_SECRET: string;
    NODE_ENV: "development" | "test" | "production";
    PORT: number;
    FRONTEND_ORIGIN: string;
    PBKDF2_ITERATIONS: number;
    DB_POOL_MAX: number;
    LOG_LEVEL: "error" | "fatal" | "warn" | "info" | "debug" | "trace" | "silent";
    REQUEST_TIMEOUT_MS: number;
};
//# sourceMappingURL=config.d.ts.map